pyproject.toml
[project]
name = "top-scraper"
version = "0.1.0"
description = "CLI tool for enriching company data from Google Sheets"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "gspread>=6.0.0",
    "httpx>=0.27.0",
    "loguru>=0.7.2",
    "openai>=1.46.0",
    "pandas>=2.2.3",
    "playwright>=1.47.0",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.1",
    "tiktoken>=0.7.0",
    "tenacity>=9.0.0",
    "tqdm>=4.66.5",
]

[project.optional-dependencies]
news = ["newsapi-python>=0.2.7"]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

